requirements.pip.txt
aiohttp==3.7.4.post0
alabaster==0.7.12
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
docutils==0.17.1
doit==0.33.1
hat-aio==0.4.1.dev20210601
hat-chatter==0.4.1.dev20210601
hat-event==0.4.1.dev20210601
hat-gateway==0.4.1.dev20210601
hat-gui==0.4.1.dev20210601
hat-json==0.4.1.dev20210601
hat-juggler==0.4.1.dev20210601
hat-monitor==0.4.1.dev20210601
hat-orchestrator==0.4.1.dev20210601
hat-peg==0.4.1.dev20210601
hat-sbs==0.4.1.dev20210601
hat-sqlite3==0.4.1.dev20210601
hat-util==0.4.1.dev20210601
idna==2.10
imagesize==1.2.0
iniconfig==1.1.1
Jinja2==3.0.1
jsonpatch==1.32
jsonpointer==2.1
jsonschema==3.2.0
lmdb==1.2.1
MarkupSafe==2.0.1
multidict==5.1.0
packaging==20.9
pluggy==0.13.1
py==1.10.0
Pygments==2.9.0
pyinotify==0.9.6
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.4
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
six==1.16.0
snowballstemmer==2.1.0
Sphinx==4.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
yarl==1.6.3

attrs==21.2.0
cffi==1.14.5
click==8.0.1
cryptography==3.4.7
decorator==4.4.2
et-xmlfile==1.1.0
flake8==3.9.2
hat-aio==0.4.1.dev20210601
hat-asn1==0.4.1.dev20210601
hat-drivers==0.4.1.dev20210601
hat-json==0.4.1.dev20210601
hat-peg==0.4.1.dev20210601
hat-util==0.4.1.dev20210601
jsonpatch==1.32
jsonpointer==2.1
jsonschema==3.2.0
llvmlite==0.36.0
mccabe==0.6.1
networkx==2.5.1
numba==0.53.1
numpy==1.21.0
openpyxl==3.0.7
packaging==20.9
pandapower==2.6.0
pandas==1.2.5
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
pyparsing==2.4.7
pyrsistent==0.17.3
pyserial==3.5
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
scipy==1.6.0
six==1.16.0
xlrd==2.0.1
XlsxWriter==1.4.3
